requirements.txt
GitPython==3.1.40
click==8.1.7
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
black==23.10.1